Add files via upload
The primary objective of GLMS is to automate the book borrowing and return processes by implementing a QR code-based system. Each book in the library is equipped with a unique QR code, and students can use their smart phones to scan the code for quick and convenient transactions the workflow begins with students scanning the QR code of a desired book. Once scanned, a ticket is generated and sent to the admin portal, where library administrators can review and approve  the request. Upon approval, the book status is updated in the student portal, indicating  successful borrowing. The system also offers features such as book renewal, providing students with the option to extend the borrowing period if needed .For returning books, a similar process is followed. Student need to click on the return button with respect to the book. A return ticket is generated and sent to the admin portal, where 
administrators can verify the return and update the book status in the student portal accordingly